Remove imports from conftest.
